helptexts and data for contact info
